+++ HISTORY.md
@@ -1,7 +1,8 @@
 ## Changelog ##
 
-### 2014-06-14 - v5.1 - (Dev time: 1 hours) ###
+### 2014-06-14 - v5.1 - (Dev time: 2 hours) ###
 - Tweak: Added `EMPTY_TRASH_DAYS` to system info page (Issue #67)
+- Tweak: Change the contextual help content for admin screens (Issue #68)
 - Bug: Added compatibility for PHP version 5.2.4 (Issue #66)
 
 ### 2014-06-12 - v5.0.2 - (Dev time: 1 hours) ###